requirements.txt
Flask==1.1.1
Flask-Login==0.5.0
google-cloud-bigquery==1.24.0
google-cloud-storage==1.25.0
google-cloud-pubsub==1.3.0
sendgrid==6.1.1
pg8000==1.13.2
psycopg2==2.8.4
SQLAlchemy==1.3.13
matplotlib==3.1.2
pandas==1.0.0
Flask_SQLAlchemy==2.4.1
Flask_Bcrypt==0.7.1
Flask_WTF==0.14.3